refactor: move `flush` helper function to `shared` package
